// src/models/models.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Time: any;
  Upload: any;
};
// ...
export type Amendment = {
  __typename?: "Amendment";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  note: Scalars["String"];
  createdAt: Scalars["Time"];
};
// ...
export type ChiefComplaint = {
  __typename?: "ChiefComplaint";
  id: Scalars["ID"];
  title: Scalars["String"];
  patientChartId: Scalars["ID"];
  hpiComponents: Array<Maybe<HpiComponent>>;
};
// ...
export type DiagnosticProcedureOrder = {
  __typename?: "DiagnosticProcedureOrder";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  patientId: Scalars["ID"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phoneNo: Scalars["String"];
  userName: Scalars["String"];
  orderedById?: Maybe<Scalars["ID"]>;
  orderedBy?: Maybe<User>;
  diagnosticProcedures: Array<DiagnosticProcedure>;
  status: DiagnosticProcedureOrderStatus;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export type EyewearPrescriptionOrder = {
  __typename?: "EyewearPrescriptionOrder";
  id: Scalars["ID"];
  eyewearShopId: Scalars["ID"];
  eyewearShop: EyewearShop;
  patientChartId: Scalars["ID"];
  orderedById?: Maybe<Scalars["ID"]>;
  orderedBy?: Maybe<User>;
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phoneNo: Scalars["String"];
  userName: Scalars["String"];
  status: Scalars["String"];
  eyewearPrescriptions: Array<Maybe<EyewearPrescription>>;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export type FollowUpOrder = {
  __typename?: "FollowUpOrder";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  patientId: Scalars["ID"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phoneNo: Scalars["String"];
  userName: Scalars["String"];
  orderedById?: Maybe<Scalars["ID"]>;
  orderedBy?: Maybe<User>;
  followUps: Array<FollowUp>;
  status: FollowUpOrderStatus;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export type LabOrder = {
  __typename?: "LabOrder";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  patientId: Scalars["ID"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phoneNo: Scalars["String"];
  userName: Scalars["String"];
  orderedById?: Maybe<Scalars["ID"]>;
  orderedBy?: Maybe<User>;
  labs: Array<Lab>;
  status: LabOrderStatus;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export type MedicalPrescriptionOrder = {
  __typename?: "MedicalPrescriptionOrder";
  id: Scalars["ID"];
  pharmacyId: Scalars["ID"];
  pharmacy: Pharmacy;
  patientChartId: Scalars["ID"];
  orderedById?: Maybe<Scalars["ID"]>;
  orderedBy?: Maybe<User>;
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phoneNo: Scalars["String"];
  userName: Scalars["String"];
  status: Scalars["String"];
  medicalPrescriptions: Array<Maybe<MedicalPrescription>>;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export type MutationDeleteOrderArgs = {
  id: Scalars["ID"];
};
// ...
export type OpthalmologyExam = {
  __typename?: "OpthalmologyExam";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  rightOrbits?: Maybe<Scalars["String"]>;
  leftOrbits?: Maybe<Scalars["String"]>;
  rightLids?: Maybe<Scalars["String"]>;
  leftLids?: Maybe<Scalars["String"]>;
  rightLacrimalSystem?: Maybe<Scalars["String"]>;
  leftLacrimalSystem?: Maybe<Scalars["String"]>;
  externalExamNote?: Maybe<Scalars["String"]>;
  rightCoverTest?: Maybe<Scalars["String"]>;
  leftCoverTest?: Maybe<Scalars["String"]>;
  coverTestNote?: Maybe<Scalars["String"]>;
  rightRetina?: Maybe<Scalars["String"]>;
  leftRetina?: Maybe<Scalars["String"]>;
  leftRetinaSketch?: Maybe<Scalars["String"]>;
  rightRetinaSketch?: Maybe<Scalars["String"]>;
  funduscopyNote?: Maybe<Scalars["String"]>;
  rightOcularMotility?: Maybe<Scalars["String"]>;
  leftOcularMotility?: Maybe<Scalars["String"]>;
  rsr?: Maybe<Scalars["String"]>;
  rio?: Maybe<Scalars["String"]>;
  rlr?: Maybe<Scalars["String"]>;
  rmr?: Maybe<Scalars["String"]>;
  rir?: Maybe<Scalars["String"]>;
  rso?: Maybe<Scalars["String"]>;
  rightFlick?: Maybe<Scalars["Boolean"]>;
  lsr?: Maybe<Scalars["String"]>;
  lio?: Maybe<Scalars["String"]>;
  llr?: Maybe<Scalars["String"]>;
  lmr?: Maybe<Scalars["String"]>;
  lir?: Maybe<Scalars["String"]>;
  lso?: Maybe<Scalars["String"]>;
  leftFlick?: Maybe<Scalars["Boolean"]>;
  distance?: Maybe<Scalars["String"]>;
  near?: Maybe<Scalars["String"]>;
  ocularMotilityNote?: Maybe<Scalars["String"]>;
  rightOpticDisc?: Maybe<Scalars["String"]>;
  leftOpticDisc?: Maybe<Scalars["String"]>;
  rightOpticDiscSketch?: Maybe<Scalars["String"]>;
  leftOpticDiscSketch?: Maybe<Scalars["String"]>;
  rightCdr?: Maybe<Scalars["String"]>;
  leftCdr?: Maybe<Scalars["String"]>;
  opticDiscNote?: Maybe<Scalars["String"]>;
  rightPupils?: Maybe<Scalars["String"]>;
  leftPupils?: Maybe<Scalars["String"]>;
  pupilsNote?: Maybe<Scalars["String"]>;
  rightConjunctiva?: Maybe<Scalars["String"]>;
  leftConjunctiva?: Maybe<Scalars["String"]>;
  rightCornea?: Maybe<Scalars["String"]>;
  leftCornea?: Maybe<Scalars["String"]>;
  rightCorneaSketch?: Maybe<Scalars["String"]>;
  leftCorneaSketch?: Maybe<Scalars["String"]>;
  leftSclera?: Maybe<Scalars["String"]>;
  rightSclera?: Maybe<Scalars["String"]>;
  rightAnteriorChamber?: Maybe<Scalars["String"]>;
  leftAnteriorChamber?: Maybe<Scalars["String"]>;
  rightIris?: Maybe<Scalars["String"]>;
  leftIris?: Maybe<Scalars["String"]>;
  rightLens?: Maybe<Scalars["String"]>;
  leftLens?: Maybe<Scalars["String"]>;
  rightLensSketch?: Maybe<Scalars["String"]>;
  leftLensSketch?: Maybe<Scalars["String"]>;
  rightVitreos?: Maybe<Scalars["String"]>;
  leftVitreos?: Maybe<Scalars["String"]>;
  slitLampExamNote?: Maybe<Scalars["String"]>;
};
// ...
export type PatientChart = {
  __typename?: "PatientChart";
  id: Scalars["ID"];
  appointmentId: Scalars["ID"];
  vitalSigns: VitalSigns;
  opthalmologyExam: OpthalmologyExam;
  physicalExamFindings: Array<PhysicalExamFinding>;
  surgicalProcedure: SurgicalProcedure;
  treatment: Treatment;
  chiefComplaints: Array<Maybe<ChiefComplaint>>;
  chiefComplaintsNote?: Maybe<Scalars["String"]>;
  bloodPressure?: Maybe<Scalars["String"]>;
  hpiNote?: Maybe<Scalars["String"]>;
  diagnosisNote?: Maybe<Scalars["String"]>;
  differentialDiagnosisNote?: Maybe<Scalars["String"]>;
  rightSummarySketch?: Maybe<Scalars["String"]>;
  leftSummarySketch?: Maybe<Scalars["String"]>;
  summaryNote?: Maybe<Scalars["String"]>;
  stickieNote?: Maybe<Scalars["String"]>;
  medicalRecommendation?: Maybe<Scalars["String"]>;
  sickLeave?: Maybe<Scalars["String"]>;
  medicalPrescriptionOrder?: Maybe<MedicalPrescriptionOrder>;
  eyewearPrescriptionOrder?: Maybe<EyewearPrescriptionOrder>;
  diagnosticProcedureOrder?: Maybe<DiagnosticProcedureOrder>;
  labOrder?: Maybe<LabOrder>;
  surgicalOrder?: Maybe<SurgicalOrder>;
  treatmentOrder?: Maybe<TreatmentOrder>;
  referralOrder?: Maybe<ReferralOrder>;
  followUpOrder?: Maybe<FollowUpOrder>;
  diagnoses: Array<Maybe<PatientDiagnosis>>;
  locked?: Maybe<Scalars["Boolean"]>;
  lockedDate?: Maybe<Scalars["Time"]>;
  lockedById?: Maybe<Scalars["ID"]>;
  lockedBy?: Maybe<User>;
  amendments: Array<Amendment>;
};
// ...
export type PhysicalExamFinding = {
  __typename?: "PhysicalExamFinding";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  examCategoryId: Scalars["ID"];
  examCategory: ExamCategory;
  abnormal: Scalars["Boolean"];
  note?: Maybe<Scalars["String"]>;
};
// ...
export type QueryTreatmentOrderArgs = {
  patientChartId: Scalars["ID"];
};
// ...
export type ReferralOrder = {
  __typename?: "ReferralOrder";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  patientId: Scalars["ID"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phoneNo: Scalars["String"];
  userName: Scalars["String"];
  orderedById?: Maybe<Scalars["ID"]>;
  orderedBy?: Maybe<User>;
  referrals: Array<Referral>;
  status: ReferralOrderStatus;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export type SurgicalOrder = {
  __typename?: "SurgicalOrder";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  patientId: Scalars["ID"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phoneNo: Scalars["String"];
  userName: Scalars["String"];
  orderedById?: Maybe<Scalars["ID"]>;
  orderedBy?: Maybe<User>;
  surgicalProcedures: Array<SurgicalProcedure>;
  status: SurgicalOrderStatus;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export type Treatment = {
  __typename?: "Treatment";
  id: Scalars["ID"];
  treatmentOrderId: Scalars["ID"];
  patientChartId: Scalars["ID"];
  note?: Maybe<Scalars["String"]>;
  result?: Maybe<Scalars["String"]>;
  rightEyeText?: Maybe<Scalars["String"]>;
  leftEyeText?: Maybe<Scalars["String"]>;
  generalText?: Maybe<Scalars["String"]>;
  treatmentTypeId: Scalars["ID"];
  treatmentType: TreatmentType;
  treatmentTypeTitle: Scalars["String"];
  status: TreatmentStatus;
  receptionNote: Scalars["String"];
  payments: Array<Payment>;
};
// ...
export type TreatmentOrder = {
  __typename?: "TreatmentOrder";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  patientId: Scalars["ID"];
  firstName: Scalars["String"];
  lastName: Scalars["String"];
  phoneNo: Scalars["String"];
  userName: Scalars["String"];
  orderedById?: Maybe<Scalars["ID"]>;
  orderedBy?: Maybe<User>;
  treatments: Array<Treatment>;
  status: TreatmentOrderStatus;
  createdAt?: Maybe<Scalars["Time"]>;
};
// ...
export enum TreatmentStatus {
  Ordered = "ORDERED",
  Completed = "COMPLETED",
}

export type TreatmentType = {
  __typename?: "TreatmentType";
  id: Scalars["ID"];
  title: Scalars["String"];
  active: Scalars["Boolean"];
  billings: Array<Maybe<Billing>>;
  supplies: Array<Maybe<Supply>>;
};
// ...
export type VitalSigns = {
  __typename?: "VitalSigns";
  id: Scalars["ID"];
  patientChartId: Scalars["ID"];
  temperature?: Maybe<Scalars["Float"]>;
  pulse?: Maybe<Scalars["Float"]>;
  bloodPressureSystolic?: Maybe<Scalars["Float"]>;
  bloodPressureDiastolic?: Maybe<Scalars["Float"]>;
  respiratoryRate?: Maybe<Scalars["Float"]>;
  oxygenSaturation?: Maybe<Scalars["Float"]>;
  height?: Maybe<Scalars["Float"]>;
  weight?: Maybe<Scalars["Float"]>;
  bmi?: Maybe<Scalars["Float"]>;
  rightDistanceUncorrected?: Maybe<Scalars["String"]>;
  leftDistanceUncorrected?: Maybe<Scalars["String"]>;
  rightDistancePinhole?: Maybe<Scalars["String"]>;
  leftDistancePinhole?: Maybe<Scalars["String"]>;
  rightDistanceCorrected?: Maybe<Scalars["String"]>;
  leftDistanceCorrected?: Maybe<Scalars["String"]>;
  rightNearUncorrected?: Maybe<Scalars["String"]>;
  leftNearUncorrected?: Maybe<Scalars["String"]>;
  rightNearPinhole?: Maybe<Scalars["String"]>;
  leftNearPinhole?: Maybe<Scalars["String"]>;
  rightNearCorrected?: Maybe<Scalars["String"]>;
  leftNearCorrected?: Maybe<Scalars["String"]>;
  rightApplanation?: Maybe<Scalars["String"]>;
  leftApplanation?: Maybe<Scalars["String"]>;
  rightTonopen?: Maybe<Scalars["String"]>;
  leftTonopen?: Maybe<Scalars["String"]>;
  rightDigital?: Maybe<Scalars["String"]>;
  leftDigital?: Maybe<Scalars["String"]>;
  rightNoncontact?: Maybe<Scalars["String"]>;
  leftNoncontact?: Maybe<Scalars["String"]>;
  rightDistanceSph?: Maybe<Scalars["String"]>;
  leftDistanceSph?: Maybe<Scalars["String"]>;
  rightDistanceAxis?: Maybe<Scalars["String"]>;
  leftDistanceAxis?: Maybe<Scalars["String"]>;
  rightDistanceCyl?: Maybe<Scalars["String"]>;
  leftDistanceCyl?: Maybe<Scalars["String"]>;
  rightNearSph?: Maybe<Scalars["String"]>;
  leftNearSph?: Maybe<Scalars["String"]>;
  rightNearCyl?: Maybe<Scalars["String"]>;
  leftNearCyl?: Maybe<Scalars["String"]>;
  rightNearAxis?: Maybe<Scalars["String"]>;
  leftNearAxis?: Maybe<Scalars["String"]>;
  rightLensMeterSph?: Maybe<Scalars["String"]>;
  leftLensMeterSph?: Maybe<Scalars["String"]>;
  rightLensMeterAxis?: Maybe<Scalars["String"]>;
  leftLensMeterAxis?: Maybe<Scalars["String"]>;
  rightLensMeterCyl?: Maybe<Scalars["String"]>;
  leftLensMeterCyl?: Maybe<Scalars["String"]>;
};